% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{deg2rad}
\alias{deg2rad}
\title{degrees to radians function}
\usage{
deg2rad(ldeg)
}
\arguments{
\item{ldeg}{Input decimal degrees. No default.}
}
\description{
Function converts decimal degrees to radians
}
\examples{
deg2rad(32.1)
}
\keyword{degrees}

