% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{doy2delta}
\alias{doy2delta}
\title{Solar declination}
\usage{
doy2delta(DOY)
}
\arguments{
\item{DOY}{day of year. No default.}
}
\description{
Function computes solar declination. Input day of year. Output radians.
}
\examples{
doy2delta(180)
}
\keyword{declination}

