% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{doy2dr}
\alias{doy2dr}
\title{Compute earth-sun distance factor}
\usage{
doy2dr(DOY)
}
\arguments{
\item{DOY}{day of year. No default.}
}
\description{
Function calculates earth-sun distance factor. Input is day of year. Output is numeric factor dimensionless.
}
\examples{
doy2dr(180)
}
\keyword{earth-sun}

