% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{elev2press}
\alias{elev2press}
\title{convert elevation to standard pressure}
\usage{
elev2press(z)
}
\arguments{
\item{uz}{Wind speed in m/s. No default.}

\item{zw}{Wind speed measurement height in m. No default.}
}
\description{
Function converts site elevation to standardized pressure
}
\examples{
getu2f(1.5,3.0)
}
\keyword{windspeed}

