% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{ETszf}
\alias{ETszf}
\title{Standardized ET}
\usage{
ETszf(Rn, G = 0, T, u2, es, ea, Delta, gamma, Cn, Cd)
}
\arguments{
\item{Rn}{Net radiation MJ m^-2 d^-1. No default.}

\item{G}{soil heat flux density MJ m^-2 d^-1. Default=0.0}

\item{T}{air temperature celsius}

\item{u2}{wind speed at 2 m heigh m}

\item{es}{saturation vapor pressure kPa}

\item{ea}{mean actual vapor pressure kPa}

\item{Delta}{slope of saturation vapor pressure curve kPa C^-1}

\item{gamma}{psychrometric constant kPa C^-1}

\item{Cn}{numerator constant 900 for daily short crop 1600 for daily tall crop}

\item{Cd}{denominator constant 0.34 for daily short crop 0.38 for daily tall crop}
}
\description{
Function computes ASCE Standardized ET
}
\examples{
ETszf(25.0,0.0,32.0,1.3,1.7,0.9,0.9,0.4,900,0.34)
}
\keyword{ETsz}

