% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{getfcdf}
\alias{getfcdf}
\title{Cloudiness fraction}
\usage{
getfcdf(Rs, Rso)
}
\arguments{
\item{Rs}{solar radiation at surface MJ m^-2 d^-1. No default.}

\item{Rso}{clear-sky radiation ata surface MJ m^-2 d^-1. No default.}
}
\description{
Function computes cloudiness fraction function. Input Rs measured or calculated solar radiation, Rso calculated clear-sky radition. Output fc cloudiness function dimensionless
}
\examples{
getfcdf(23.0,25.0)
}
\keyword{cloudiness}

