% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{getRnlf}
\alias{getRnlf}
\title{Net Long-Wave Radiation}
\usage{
getRnlf(fcd, ea, tkmax, tkmin)
}
\arguments{
\item{fcd}{Cloudiness fraction function dimensionless. No default.}

\item{ea}{Daily average vapor pressure kPa. No default.}

\item{tkmax}{Daily maximum air temperature K. No default.}

\item{tkmin}{Daily minimum air temperature K. No default.}
}
\description{
Function computes net longwave radiation from cloudiness fraction, vapor pressure, daily air temperature range. Input fcd from getfcdf function, ea vapor pressure kPa, daily maximum and minimum air temperature in kelvin. Output Rnl MJ m^-2 d^-1
}
\examples{
getRnlf(0.1,1.5,300.0,291.0)
}
\keyword{Longwave}

