% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{getu2f}
\alias{getu2f}
\title{convert wind speed to 2m value function}
\usage{
getu2f(uz, zw)
}
\arguments{
\item{uz}{Wind speed in m/s. No default.}

\item{zw}{Wind speed measurement height in m. No default.}
}
\description{
Function converts wind speed at measuremnt height to equivalent wind speed at 2m above the ground level
}
\examples{
getu2f(1.5,3.0)
}
\keyword{windspeed}

