% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{press2psychro}
\alias{press2psychro}
\title{convert site pressure to psychrometric constant}
\usage{
press2psychro(P)
}
\arguments{
\item{P}{pressure in kPa. No default.}
}
\description{
Function converts site pressure to standardized psychrometric constant. Output in kPa.
}
\examples{
press2psychro(98.1)
}
\keyword{psychrometric}

