% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{Ra2Rsof}
\alias{Ra2Rsof}
\title{Clear Sky radiation}
\usage{
Ra2Rsof(Ra, stationelevm)
}
\arguments{
\item{Ra}{extraterrestrial radiation MJ m^-2 d^-1. No default.}

\item{stationelevm}{meters. No default.}
}
\description{
Function computes clear sky radiation at surface. Input extraterrestrial radiation obtained from Radailyf and station elevation in meters. Output Rso in MJ m^-2 d^-1
}
\examples{
Ra2Rsof(23.0,312.2)
}
\keyword{clear}
\keyword{sky}

