% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{rad2deg}
\alias{rad2deg}
\title{radians to degrees function}
\usage{
rad2deg(lrad)
}
\arguments{
\item{lrad}{Input radians. No default.}
}
\description{
Function converts radians to decimal degrees
}
\examples{
rad2deg(pi/2.0)
}
\keyword{radians}

