% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{Radailyf}
\alias{Radailyf}
\title{Extraterrestrial radiation}
\usage{
Radailyf(dr, omegas, sitelatdeg, delta)
}
\arguments{
\item{dr}{earth-sun distance factor, inverse squared, dimensionless. No default.}

\item{omegas}{sunset hour angle radians, computed by sunsethourangf}

\item{sitelatdeg}{site location latitude decimal degrees}

\item{delta}{solar declination radians. Compute using doy2delta function.}
}
\description{
Function computes daily extraterrestrial radiation. Input earth-sun distance factor, sunset hour angle, site latitude in degrees, solar declination in radians. Output Ra in MJ m^-2 d^-1
}
\examples{
Radailyf(0.95,1.5,32.2.,0.6)
}
\keyword{extraterrestrial}

