% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{Rsolar2Rns}
\alias{Rsolar2Rns}
\title{Compute net shortwave radiation from incoming solar radiation}
\usage{
Rsolar2Rns(Rs, albedo = 0.23)
}
\arguments{
\item{Rs}{incoming solar radiation MJ m^-2 d^-1. No default.}

\item{albedo}{is surface albedo dimensionless. Default=0.23}
}
\description{
Function calculates daily net shortwave radiation given incoming solar radiation. Input is MJ m^-2 d^-1. Output is MJ m^-2 d^-1.
}
\examples{
Rsolar2Rns(29.6,0.22)
}
\keyword{esat}

