% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{sunsethourangf}
\alias{sunsethourangf}
\title{Sunset hour angle}
\usage{
sunsethourangf(sitelatdeg, delta)
}
\arguments{
\item{sitelatdeg}{latitude degrees. No default.}

\item{delta}{solar declination radians. Compute using doy2delta function.}
}
\description{
Function sunset hour angle from site latitude and solar declination. Input latitude in degrees but declination in radians. Output omega_s hours.
}
\examples{
sunsethourangf(32.1,0.6)
}
\keyword{sunset}

