% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{tc2tk}
\alias{tc2tk}
\title{Celsius to Kelvin}
\usage{
tc2tk(tc)
}
\arguments{
\item{tc}{temperature in celsius. No default.}
}
\description{
Function converts celsius to kelvin
}
\examples{
tc2tk(27.0)
}
\keyword{celsius}

