% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{tcair2Delta}
\alias{tcair2Delta}
\title{Compute slope of saturation vapor pressure curve}
\usage{
tcair2Delta(TC)
}
\arguments{
\item{TC}{air temperature in celsius. No default.}
}
\description{
Function calculates the slope of the saturation vapor pressure curve at the given air temperature. Input is Celsius. Output is kPa/C.
}
\examples{
tcair2Delta(25.1)
}
\keyword{Delta}

