% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{tcair2esat}
\alias{tcair2esat}
\title{Compute saturation vapor pressure for H2O}
\usage{
tcair2esat(TC)
}
\arguments{
\item{TC}{air temperature in celsius. No default.}
}
\description{
Function calculates saturation vapor pressure at the given air temperature. Input is Celsius. Output is kPa.
}
\examples{
tcair2esat(25.1)
}
\keyword{esat}

