% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/et_functions.R
\name{tk2tc}
\alias{tk2tc}
\title{Kelvin to Celsius}
\usage{
tk2tc(tk)
}
\arguments{
\item{tk}{temperature in kelvin. No default.}
}
\description{
Function converts kelvin to celsius
}
\examples{
tk2tc(300.0)
}
\keyword{kelvin}

